requirements.txt
PyPDF2==3.0.1
python-docx==0.8.11
reportlab==4.0.4
tkinterdnd2==0.3.0